WIP: Organizing Cypress UI Tests
