+++ test/webclient/cypress/support/commands.js
@@ -1,19 +1,24 @@
 "use strict";
 
-import * as selectors from '../support/utils/get_selectors';
 import * as constants from '../support/utils/constants';
 
 // ***********************************************
 // This file contains custom build command for cypress
 // ***********************************************
 
-// -- This is a custom command to login into AIR application --
-Cypress.Commands.add("login", function(){
+Cypress.Commands.add("login", function () {
+
+    /*
+    * Clear cookies before login into application. Because We are preserving cookies to use it all test suit.
+    * Running test suit multiple times require new session to login into application.
+    */
+    cy.clearCookie(constants.APPLICATION_COOKIE);
+
 
     /************************************
      * Login into application
-     * 1. Fill username and password
-     * 2. Click Login button
+     * 1. Send request to endpoint: /v1/authn
+     * 2. Pass the necessary parameters
      * ***********************************/
 
     let username;
@@ -27,7 +32,7 @@ Cypress.Commands.add("login", function(){
         return {"user": config.Tester1Name, "pass": config.Tester1Pass};
     }).then((creds) => {
 
-        username    = creds.user;
+        username = creds.user;
         password = creds.pass;
 
         // log format
@@ -43,44 +48,25 @@ Cypress.Commands.add("login", function(){
         });
 
         // login steps
-        cy
-            .get('input[name=user]').type(username).should('have.value', username) // enter username
-            .get('input[name=pass]').type(password).should('have.value', password) // enter password
-            .get('button[name=login]').click().wait(2000) // click on login and wait for 1s to get the dashboard page
+        cy.request('POST', '/v1/authn/', {"user": username, "pass": password})
+            .then((resp) => {
+                cy.log(resp);
+                expect(resp.status).to.eq(200)
+            });
+
+        // Check Cookie exists after login into application
+        cy.getCookie(constants.APPLICATION_COOKIE).should('exist')
             .then(function () {
-                log.snapshot().end(); //end custom command
+                log.snapshot().end();
             });
     });
 
 });
 
-// -- This is a custom command to Change Business Unit to REX --
-Cypress.Commands.add("changeBU", function (appSettings) {
-    let log;
-
-    // log format
-    log = Cypress.log({
-        name: "Change Business to REX",
-        message: [constants.testBiz, constants.testBizID],
-    });
-
-    // get business id from appSettings variable for 'REX'
-    appSettings.BizMap.forEach(function (item) {
-        if (item.BUD === constants.testBiz) {
-            constants.testBizID = item.BID;
-        }
-    });
-
-    // Now change the business to REX
-    cy.get('[name="BusinessSelect"]').select(constants.testBiz);
-
-    // Check BusinessSelect value is set per the expected BID from appSettings variable
-    cy.get('[name="BusinessSelect"]').should('have.value', constants.testBizID.toString());
-
-    // onSuccessful test set BID value. If above test get fail below code will not be executed.
-    constants.BID = constants.testBizID;
-
-    cy.log("Business ID: ", constants.BID);
-
-    log.snapshot().end(); //end custom command
+Cypress.Commands.add("logout", function () {
+    // This command log off the application
+    cy.request('GET', '/v1/logoff/')
+        .then((resp) => {
+        expect(resp.status).to.eq(200);
+        });
 });
