+++ test/webclient/cypress/support/defaults.js
@@ -1,0 +1,3 @@
+import * as constants from './utils/constants';
+
+Cypress.Cookies.defaults({whitelist: constants.APPLICATION_COOKIE});
