+++ test/webclient/cypress/support/index.js
@@ -17,6 +17,7 @@
 
 // Import commands.js using ES2015 syntax:
 import './commands'
+import './defaults'
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
